+++ src/modules/users/dto/user.dto.ts
@@ -1,0 +1,17 @@
+import { IsNotEmpty } from 'class-validator';
+
+export class CreateUserDto {
+  @IsNotEmpty()
+  login: string;
+
+  @IsNotEmpty()
+  password: string;
+}
+
+export class UpdatePasswordDto {
+  @IsNotEmpty()
+  oldPassword: string;
+
+  @IsNotEmpty()
+  newPassword: string;
+}
