--- src/utils/dto.ts
@@ -1,32 +1,0 @@
-import { IsNotEmpty, IsUUID } from 'class-validator';
-import { Track } from './types';
-
-export class CreateUserDto {
-  @IsNotEmpty()
-  login: string;
-
-  @IsNotEmpty()
-  password: string;
-}
-
-export class UpdatePasswordDto {
-  @IsNotEmpty()
-  oldPassword: string;
-
-  @IsNotEmpty()
-  newPassword: string;
-}
-
-export class TrackDto implements Track {
-  @IsNotEmpty()
-  name: string;
-
-  @IsNotEmpty()
-  artistId: string | null;
-
-  @IsNotEmpty()
-  albumId: string | null;
-
-  @IsNotEmpty()
-  duration: number;
-}
